api_service/requirements.txt
fastapi
httpx
uvicorn
grpcio
grpcio-tools
python-jose
passlib
